Added administration controller to provide a page for managing articles
Uses the RESTful API

+++ WebShop/WebShop/Controllers/ArticlesController.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
+using System.Net.Http;
 using System.Web.Http;
 
 namespace WebShop.Controllers
@@ -20,6 +22,11 @@ namespace WebShop.Controllers
             return (int)Math.Ceiling((double)_repository.Count() / _pageSize);
         }
 
+        public IEnumerable<Article> GetArticles()
+        {
+            return _repository.Get();
+        }
+
         public IEnumerable<Article> GetArticles(int page)
         {
             return _repository.Get().Skip((page > 0 ? page - 1 : 0) * _pageSize).Take(_pageSize);
@@ -27,9 +34,15 @@ namespace WebShop.Controllers
 
         public Article Get(int id)
         {
-            Article article;
-            _repository.TryGet(id, out article);
+            var article = _repository.Get(id);
+            if (article == null)
+                throw new HttpResponseException(Request.CreateResponse(HttpStatusCode.NotFound));
             return article;
+        }
+
+        public void Put(int id, [FromBody]Article article)
+        {
+            _repository.Add(article);
         }
 
         public void Post([FromBody]Article article)
@@ -39,6 +52,9 @@ namespace WebShop.Controllers
 
         public void Delete(int id)
         {
+            var article = _repository.Get(id);
+            if (article == null)
+                throw new HttpResponseException(Request.CreateResponse(HttpStatusCode.NotFound));
             _repository.Delete(id);
         }
     }
